changed map

+++ app/src/main/java/net/faracloud/dashboard/core/model/ObservationRemoteModel.kt
@@ -3,4 +3,4 @@ package net.faracloud.dashboard.core.model
 data class ObservationRemoteModel(var value: String,
                                   var timestamp: String,
                                   var location: String,
-                                  var time: Int)
+                                  var time: Long)
